REFACTOR: Put all Pydantic models in their own files
