+++ server/api/routers/classes.py
@@ -1,6 +1,3 @@
-from pydantic import BaseModel, ConfigDict, Field
-from pydantic.types import Annotated
-
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import select
 from sqlalchemy.exc import IntegrityError
@@ -8,10 +5,9 @@ from sqlalchemy.orm import Session
 
 from server.api import get_db
 from server.logger.logger import logger
-from server.api.models.base_response import BaseResponse
-from server.api.models.class_subclass_bases import ClassBase, SubclassBase
-from server.api.models.delete_response import DeleteResponse
 from server.database.models.classes import Class
+from server.api.models.delete_response import DeleteResponse
+from server.api.models.cls import ClassModel, ClassPostBase, ClassPutBase, ClassResponse
 
 
 router = APIRouter(
@@ -19,52 +15,6 @@ router = APIRouter(
     tags=["Classes"],
     responses={404: {"description": "Not found."}},
 )
-
-
-class ClassModel(ClassBase):
-    """
-    Represents a class entity.
-
-    - `id`: Unique identifier of the class.
-    - `name`: Name of the class.
-    - `subclasses`: List of related subclass entities.
-    """
-
-    subclasses: list[SubclassBase] | None
-
-    model_config = ConfigDict(from_attributes=True)
-
-
-class ClassPostBase(BaseModel):
-    """
-    Schema for creating a new class.
-
-    - `class_name`: Name of the class to be created, must be between 1 and 50 characters.
-    """
-
-    class_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class ClassPutBase(BaseModel):
-    """
-    Schema for updating a class.
-
-    - `class_name`: Name of the class to be created, must be between 1 and 50 characters.
-    """
-
-    class_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class ClassResponse(BaseResponse):
-    """
-    Response model for creating or retrieving a class.
-    Inherits from BaseResponse
-
-    - `message`: A descriptive message about the action performed.
-    - `cls`: The actual class data, represented by the `ClassModel`.
-    """
-
-    cls: ClassModel
 
 
 @router.get("/", response_model=list[ClassModel])
