+++ server/api/routers/enemies.py
@@ -7,15 +7,6 @@ from sqlalchemy.exc import IntegrityError
 
 from server.api import get_db
 from server.logger.logger import logger
-from server.api.utils.utilities import Utilities
-from server.api.utils.creature_utilities import CreatureUtilities
-from server.api.models.base_response import BaseResponse
-from server.api.models.creatures import (
-    CreatureModel,
-    CreaturePostBase,
-    CreaturePutBase,
-)
-from server.api.models.delete_response import DeleteResponse
 from server.database.models.sizes import Size
 from server.database.models.types import Type
 from server.database.models.classes import Class
@@ -24,24 +15,21 @@ from server.database.models.enemies import Enemy
 from server.database.models.races import Race
 from server.database.models.subraces import Subrace
 from server.database.models.parties import Party
+from server.api.utils.utilities import Utilities
+from server.api.utils.creature_utilities import CreatureUtilities
+from server.api.models.delete_response import DeleteResponse
+from server.api.models.enemy import EnemyResponse
+from server.api.models.creatures import (
+    CreatureModel,
+    CreaturePostBase,
+    CreaturePutBase,
+)
 
 router = APIRouter(
     prefix="/api/enemies",
     tags=["Enemies"],
     responses={404: {"description": "Not found."}},
 )
-
-
-class EnemyResponse(BaseResponse):
-    """
-    Response model for creating or retrieving a enemy.
-    Inherits from BaseResponse
-
-    - `message`: A descriptive message about the action performed.
-    - `enemy`: The actual enemy data, represented by the `CreatureModel`.
-    """
-
-    enemy: CreatureModel
 
 
 @router.get("/", response_model=list[CreatureModel])
