+++ server/api/routers/player_characters.py
@@ -1,19 +1,12 @@
 from typing import Any
 
 from fastapi import APIRouter, Depends, HTTPException
-from pydantic import BaseModel, ConfigDict
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 from sqlalchemy.exc import IntegrityError
 
 from server.api import get_db
 from server.logger.logger import logger
-from server.api.utils.creature_utilities import CreatureUtilities
-from server.api.utils.utilities import Utilities
-from server.api.models.delete_response import DeleteResponse
-from server.api.models.base_response import BaseResponse
-from server.api.models.creatures import CreatureModel, CreaturePostBase, CreaturePutBase
-from server.api.models.user_relations import UserBase
 from server.database.models.sizes import Size
 from server.database.models.types import Type
 from server.database.models.classes import Class
@@ -22,47 +15,17 @@ from server.database.models.player_characters import PlayerCharacter
 from server.database.models.races import Race
 from server.database.models.subraces import Subrace
 from server.database.models.parties import Party
+from server.api.models.delete_response import DeleteResponse
+from server.api.models.creatures import CreaturePutBase
+from server.api.models.player_character import PCModel, PCPostBase, PCResponse
+from server.api.utils.creature_utilities import CreatureUtilities
+from server.api.utils.utilities import Utilities
 
 router = APIRouter(
     prefix="/api/player_characters",
     tags=["Player Characters"],
     responses={404: {"description": "Not found."}},
 )
-
-
-class PCModel(CreatureModel):
-    user: UserBase
-
-
-class PCPostBase(CreaturePostBase):
-    """
-    Extension of the CreaturePostBase with extra user_id for
-    pc to user connection.
-    """
-
-    user_id: int
-
-
-class UserPublic(BaseModel):
-    """Only allows specific data from the User table to show up."""
-
-    name: str
-    username: str
-    image: bytes | None = None
-
-    model_config = ConfigDict(from_attributes=True)
-
-
-class PCResponse(BaseResponse):
-    """
-    Response model for creating or retrieving a player character.
-    Inherits from BaseResponse
-
-    - `message`: A descriptive message about the action performed.
-    - `pc`: The actual pc data, represented by the `PCModel`.
-    """
-
-    pc: PCModel
 
 
 @router.get("/", response_model=list[PCModel])
