+++ server/api/routers/roles.py
@@ -1,6 +1,3 @@
-from pydantic import BaseModel, Field
-from pydantic.types import Annotated
-
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import select
 from sqlalchemy.orm import Session
@@ -8,58 +5,15 @@ from sqlalchemy.exc import IntegrityError
 
 from server.api import get_db
 from server.logger.logger import logger
-from server.api.models.base_response import BaseResponse
-from server.api.models.user_relations import RoleBase
-from server.api.models.delete_response import DeleteResponse
-from server.api.routers.users import UserModel
 from server.database.models.roles import Role
+from server.api.models.delete_response import DeleteResponse
+from server.api.models.role import RoleModel, RolePostBase, RolePutBase, RoleResponse
 
 router = APIRouter(
     prefix="/api/roles",
     tags=["Roles"],
     responses={404: {"description": "Not found."}},
 )
-
-
-class RoleModel(RoleBase):
-    """
-    Extension of the RoleBase entity.
-
-    - `users`: List of all users that have this role.
-    """
-
-    users: list[UserModel] | None
-
-
-class RolePostBase(BaseModel):
-    """
-    Schema for creating a new role.
-
-    - `role_name`: Name of the role to be created, must be between 1 and 50 characters.
-    """
-
-    role_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class RolePutBase(BaseModel):
-    """
-    Schema for updating an role.
-
-    - `role_name`: Updated name of the role, must be between 1 and 50 characters.
-    """
-
-    role_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class RoleResponse(BaseResponse):
-    """
-    Response model for creating or retrieving a role.
-
-    - `message`: A descriptive message about the action performed.
-    - `role`: The actual role data, represented by the `RoleModel`.
-    """
-
-    role: RoleModel
 
 
 @router.get("/", response_model=list[RoleModel])
