+++ server/api/routers/sizes.py
@@ -1,6 +1,3 @@
-from pydantic import BaseModel, ConfigDict, Field
-from pydantic.types import Annotated
-
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import select
 from sqlalchemy.orm import Session
@@ -8,61 +5,15 @@ from sqlalchemy.exc import IntegrityError
 
 from server.api import get_db
 from server.logger.logger import logger
-from server.api.models.base_response import BaseResponse
-from server.api.models.delete_response import DeleteResponse
 from server.database.models.sizes import Size
+from server.api.models.size import SizeModel, SizePostBase, SizePutBase, SizeResponse
+from server.api.models.delete_response import DeleteResponse
 
 router = APIRouter(
     prefix="/api/sizes",
     tags=["Sizes"],
     responses={404: {"description": "Not found."}},
 )
-
-
-class SizeModel(BaseModel):
-    """
-    Represents a size entity.
-
-    - `id`: Unique identifier of the size.
-    - `name`: Name of the size.
-    """
-
-    id: int
-    name: str
-
-    model_config = ConfigDict(from_attributes=True)
-
-
-class SizePostBase(BaseModel):
-    """
-    Schema for creating a new size.
-
-    - `size_name`: Name of the size to be created, must be between 1 and 50 characters.
-    """
-
-    size_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class SizePutBase(BaseModel):
-    """
-    Schema for updating an size.
-
-    - `size_name`: Updated name of the size, must be between 1 and 50 characters.
-    """
-
-    size_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class SizeResponse(BaseResponse):
-    """
-    Response model for creating or retrieving a size.
-    Inherits from BaseResponse
-
-    - `message`: A descriptive message about the action performed.
-    - `size`: The actual size data, represented by the `RoleModel`.
-    """
-
-    size: SizeModel
 
 
 @router.get("/", response_model=list[SizeModel])
