+++ server/api/routers/types.py
@@ -1,6 +1,3 @@
-from pydantic import BaseModel, ConfigDict, Field
-from pydantic.types import Annotated
-
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy import select
 from sqlalchemy.orm import Session
@@ -8,9 +5,9 @@ from sqlalchemy.exc import IntegrityError
 
 from server.api import get_db
 from server.logger.logger import logger
-from server.api.models.base_response import BaseResponse
-from server.api.models.delete_response import DeleteResponse
 from server.database.models.types import Type
+from server.api.models.delete_response import DeleteResponse
+from server.api.models.type import TypeModel, TypePostBase, TypePutBase, TypeResponse
 
 
 router = APIRouter(
@@ -18,52 +15,6 @@ router = APIRouter(
     tags=["Types"],
     responses={404: {"description": "Not found."}},
 )
-
-
-class TypeModel(BaseModel):
-    """
-    Represents a type entity.
-
-    - `id`: Unique identifier of the type.
-    - `name`: Name of the type.
-    """
-
-    id: int
-    name: str
-
-    model_config = ConfigDict(from_attributes=True)
-
-
-class TypePostBase(BaseModel):
-    """
-    Schema for creating a new type.
-
-    - `type_name`: Name of the type to be created, must be between 1 and 50 characters.
-    """
-
-    type_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class TypePutBase(BaseModel):
-    """
-    Schema for updating an type.
-
-    - `type_name`: Updated name of the type, must be between 1 and 50 characters.
-    """
-
-    type_name: Annotated[str, Field(min_length=1, max_length=50)]
-
-
-class TypeResponse(BaseResponse):
-    """
-    Response model for creating or retrieving an type.
-    Inherits from BaseResponse
-
-    - `message`: A descriptive message about the action performed.
-    - `type`: The actual type data, represented by the `TypeModel`.
-    """
-
-    type: TypeModel
 
 
 @router.get("/", response_model=list[TypeModel])
